+++ src/main/java/ru/imanov/blog/service/impl/TagServiceImpl.java
@@ -4,9 +4,11 @@ import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Service;
 import ru.imanov.blog.entity.Tag;
+import ru.imanov.blog.exception.common.NullIdException;
 import ru.imanov.blog.exception.tag.TagAlreadyExistsException;
 import ru.imanov.blog.exception.tag.TagFieldsEmptyException;
 import ru.imanov.blog.exception.tag.TagNotFoundException;
+import ru.imanov.blog.exception.user.UserNotFoundException;
 import ru.imanov.blog.repository.TagRepository;
 import ru.imanov.blog.service.TagService;
 
@@ -23,11 +25,20 @@ public class TagServiceImpl implements TagService {
      * @param tag - tag to update
      * @return - updated tag
      * @throws TagNotFoundException - it is thrown out when the user is not in the database
+     * @throws NullIdException - it is thrown out when the tag has id = null
      */
     @Override
-    public Tag update(Tag tag) throws TagNotFoundException {
-        if (tagRepository.existsById(tag.getId())){
-            throw new TagNotFoundException("the tag cannot be updated because it is not found", true);
+    public Tag update(Tag tag) throws TagNotFoundException, NullIdException {
+
+        if (tag.getId() == null){
+            throw new NullIdException("you cannot update a tag with id = null");
+        }
+
+        if (!tagRepository.existsById(tag.getId())){
+            throw new TagNotFoundException(
+                    String.format("the tag with id = %d cannot be updated because it is not found", tag.getId()),
+                    true
+            );
         }
 
         checkTag(tag);
@@ -44,7 +55,9 @@ public class TagServiceImpl implements TagService {
     @Override
     public Tag add(Tag tag) throws TagAlreadyExistsException {
         if (tag.getId() != null && tagRepository.existsById(tag.getId())){
-            throw new TagAlreadyExistsException("the tag cannot be added as it already exists", true);
+            throw new TagAlreadyExistsException(
+                    String.format("the tag with id = %d cannot be added as it already exists", tag.getId()),
+                    true);
         }
 
         checkTag(tag);
@@ -63,7 +76,7 @@ public class TagServiceImpl implements TagService {
         Optional<Tag> tagFromDb = tagRepository.findById(id);
 
         if (tagFromDb.isEmpty()){
-            throw new TagNotFoundException("tag not found by id");
+            throw new TagNotFoundException(String.format("tag not found by id = %d", id));
         }
 
         return tagFromDb.get();
@@ -75,6 +88,15 @@ public class TagServiceImpl implements TagService {
      */
     @Override
     public void delete(Long id) {
+
+        if (!tagRepository.existsById(id)){
+            throw new UserNotFoundException(String.format(
+                    "a tag with id = %d cannot be deleted because he is not in the database",
+                    id
+                )
+            );
+        }
+
         tagRepository.deleteById(id);
     }
