+++ src/main/java/ru/imanov/blog/service/impl/UserServiceImpl.java
@@ -3,6 +3,7 @@ package ru.imanov.blog.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.StringUtils;
+import ru.imanov.blog.exception.common.NullIdException;
 import ru.imanov.blog.exception.user.UserAlreadyExistsException;
 import org.springframework.stereotype.Service;
 import ru.imanov.blog.entity.User;
@@ -31,11 +32,20 @@ public class UserServiceImpl implements UserService {
      * @param user - user to update
      * @return - updated user
      * @throws UserNotFoundException - it is thrown out when the user is not in the database
+     * @throws NullIdException - it is thrown out when the user has id = null
      */
     @Override
-    public User update(User user) throws UserNotFoundException{
+    public User update(User user) throws UserNotFoundException, NullIdException {
+
+        if (user.getId() == null){
+            throw new NullIdException("you cannot update a user with id = null");
+        }
+
         if (!userRepository.existsById(user.getId())){
-            throw new UserNotFoundException("the user cannot be updated because it is not found", true);
+            throw new UserNotFoundException(
+                    String.format("the user with id = %d cannot be updated because it is not found", user.getId()),
+                    true
+            );
         }
 
         checkUser(user);
@@ -50,10 +60,14 @@ public class UserServiceImpl implements UserService {
      * @return - added user
      * @throws UserAlreadyExistsException - it is thrown out when the user is already in the database
      */
+
     @Override
-    public User add(User user) throws UserAlreadyExistsException{
+    public User add(User user) throws UserAlreadyExistsException {
         if (user.getId() != null && userRepository.existsById(user.getId())){
-            throw new UserAlreadyExistsException("the user cannot be added as it already exists", true);
+            throw new UserAlreadyExistsException(
+                    String.format("the user with id = %d cannot be added as it already exists", user.getId()),
+                    true
+            );
         }
 
         checkUser(user);
@@ -69,11 +83,11 @@ public class UserServiceImpl implements UserService {
      * @throws UserNotFoundException -  it is thrown out when the user is not in the database
      */
     @Override
-    public User getById(Long id) throws UserNotFoundException{
+    public User getById(Long id) throws UserNotFoundException {
         Optional<User> userFromDb = userRepository.findById(id);
 
         if (userFromDb.isEmpty()){
-            throw new UserNotFoundException("user not found by id");
+            throw new UserNotFoundException(String.format("user not found by id = %d", id));
         }
 
         return userFromDb.get();
@@ -85,6 +99,14 @@ public class UserServiceImpl implements UserService {
      */
     @Override
     public void delete(Long id) {
+        if (!userRepository.existsById(id)){
+            throw new UserNotFoundException(String.format(
+                        "a user with id = %d cannot be deleted because he is not in the database",
+                        id
+                    )
+            );
+        }
+
         userRepository.deleteById(id);
     }
